+++ app/smartbrain_3000/web/_app/immutable/entry/start.BYUNrWQs.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/B9Y7qh5v.js";export{e as load_css,t as start};