+++ app/smartbrain_3000/web/_app/immutable/entry/start.XLc3SAPj.js
@@ -1,1 +1,0 @@
-import{a as e,n as t}from"../chunks/C-1FJjmk.js";export{e as load_css,t as start};